Crate readwrite_comp

This library purpose is to stack writers (and subsequently read with similar stack).

This allow composition of writers.

The library uses std::io::write and std::io::read for basis, but semantically Write and Read implementation requires a bit of care (see examples).

Ext trait add a possible header writing/reading and a possible end of message writing/reading.

Read implementation should return the read size, and if read return 0 it means that the reading is finished. (read ext trait should finalize the read (waiting for a new header)).

Most of this library is using WriteExt and ReadExt trait which allow to define additional action over standard read and write for example : - encyphering content : an additional header is required in most case. - adding info to content : like control, for instance an end of frame byte (required an end of message). - linking two reader or two writer (for instance CompExtW do it) The point is that WriteExt and ReadExt does not compose over the internal reader/writer to allow things such as MultiW or MultiR where we got a final Writer or final Reader but an undefined number of ExtWriter and ExtReader (and still static type without fat pointer).

WriteExt and ReadExt could be composed, using MultiW/R or CopmExtW/R.

WriteExt and ReadExt could be used as standard Reader or Writer by using CompW or CompR,

Composition by creating CompW of CompW as Writer and CompW as WriterExt is not really encouraged (even if some test are included) due to difficulty to write header or end of message recursivly (the first component is seen as a Read or a Write). CompW should in priority as a last wrapper. CompExtWInner and CompExtRInner are an alternative to CompW and CompR with less overhead but no guaranties over header and end of message. (first they was private but prove usefull in some cases).

Current issue with this crate are - drop cause panic in panic when read_end / write_end panic : leading to no clue about the issue - flush is not recursive : only the extWrite flush, the inner writer does not : TODO add a bool in flush_into to say if inner writer must be flush : that way on drop inner writer will not be flush (but still contain end_write), and have the composition flush flush all the way. Plus expose in object (CompW, MultiComp) this boolean. CompExtW is already recursive by default and should stay like that. : inner writer must flush at each layer. (MultiW flush until the inner writer of course) - flush and write_end semantic is tricky : flush means that no counterpart is needed in read (we can flush anywhere without a failure from read) whereas write_end involve a read_end need. Read_end is triggered manually with two cases : - we know the length to read (for instance a serialized object read from a reader) : that is easy and not a big issue - we do not know the length (for instance proxy of encrypted content) : then a reader like endstream could be used and read will block when seeing a end content by returning Ok(0) until read_end is used to unlock the reading. There is a serious limitition here when composing this kind of bloquing reader with non bloquing reader having some end content : the bloquing one must be in the internal layer, otherwhise the end sequence of the non bloquing will be skipped (this is a somehow tricky case but also not so common (some cipher does not require end as using flush is enough (endstream may still be used for perf when proxying content that we do not want to read)). - symetry between read and write is not enforced, non symetric implementation will fail

Example of usage could be found it tests, but also in mydht-base tunnel implementation, mydht-base bytes_wr and mydht shadow (for example mydht-openssl).

Structs

CompExtR

CompExtRInner
Inner construct to build a read upon another one, do not use as write if you need automatic header or automatic end (technical).

CompExtW
Compose two ExtWrite in a single on with Owned ExtWrite. EW1 apply over EW2 meaning that EW2 is the external layer (ew2 header written first without applying ew1 over it and ew2 end written last without ew1 written over it and content written by first applying ew2 then ew1.

CompExtWInner
inner struct for implemention just to apply method of sw in write This is not to be use directly as write but just to use write_into and flush_into

CompR
Base construct to build a read upon another one (composable reader).

CompW
Base construct to build a write upon another one (composable writer), using an ExtWrite implementation.

MultiRExt

MultiWExt

Enums

CompRState

CompWState

Traits

ExtRead
Compose over a reader with additional possibility to read an end content

ExtWrite
Write with further common functionnalities.

Functions

new_multir

new_multiw

Type Definitions

MultiR
Multiple layered read (similar to MCompW).

MultiW
CompW with several (undefined number at compile time) Write of same kind to chain, and a dest write. Drop semantic for use as write cannot be enable (inner temporary use of MCompW would write_end at each write). This could be use for layered write (for example in a multilayer ssl tunnel).