Struct readwrite_comp::CompW

pub struct CompW<'a, 'b, W: 'a + Write, EW: 'b + ExtWrite>(pub &'a mut W, pub &'b mut EW, pub CompWState);

Base construct to build a write upon another one (composable writer), using an ExtWrite implementation.

This is used in many place with short lifecycle, it would be interesting to evaluate the overhead (or see if it is optimized).

When composing multiple ExtWrite, one could compose with multiple CompW , but as the first parameter is simply Write, write end will not be called recursively (for usage as a standard Write it is safe as write end will be called recursively through drop (yet with no catching of the possible errors). For composing with full support for write_end, consider only one layer of CompW and composing the ExtWrit with a CompExtW composition.

Methods

impl<'a, 'b, W: 'a + Write, EW: 'b + ExtWrite> CompW<'a, 'b, W, EW>

fn new(w: &'a mut W, ew: &'b mut EW) -> Self

fn suspend(self) -> Result<(&'b mut EW, CompWState)>

suspend write (inner writer is available again) but keep reference for subsequent write in same state

fn resume(with: &'a mut W, from: (&'b mut EW, CompWState)) -> Self

fn write_end(&mut self) -> Result<()>