readwrite_comp::MultiW

type MultiW<'a, 'b, W, EW> = CompW<'a, 'b, W, MultiWExt<EW>>;

CompW with several (undefined number at compile time) Write of same kind to chain, and a dest write. Drop semantic for use as write cannot be enable (inner temporary use of MCompW would write_end at each write). This could be use for layered write (for example in a multilayer ssl tunnel).

Suspend (up to W) and write end will impact all layer.

Order of layer is external layer last (the writer is therefore logicaly at the end of the array of layer).